Established new API idiom for **kw/mapping/items variants of API calls.
Specifically, the new preferred API parameter form is to accept a list of
key/value pairs (e.g. 'dict.items()'), as that is generally the most
efficient format for non-trivial manipulation.  In essence, if an API
call previously took '**' keyword arguments (and perhaps a mapping as
well), it probably now takes an item list.  The exceptions are things
like constructors for infrequently created objects, and calls which pass
keyword arguments through to outside functions (e.g. in TW.Callbacks).

For situations where the old syntactic sugar is desired, there is now an
'Items()' API call which converts an optional mapping and keyword
arguments into an itmes list for you.  See the docs for the TW.API.Misc
module for details.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@345 571e12c6-e1fa-0310-aee7-ff1267fa46bd