Moved module inheritance from 'peak.api' to 'peak.binding', and reformatted peak.api package accordingly.  Applied the following name changes::
 .getSEFparent()  -> .getParentComponent()
 ._setSEFparent() -> .setParentComponent()
 .getService()    -> .lookupComponent()

 setupModule()    -> binding.setupModule()
 adviseModule()   -> binding.patchModule()
 configure()      -> binding.configure()

Also made the default '__proceed__' value a built-in, so that it's not
necessary to import it.

This concludes the essential code-level transformations from TransWarp
to PEAK.  Almost everything should be in its long-term home now, with
its long-term name, although the 'model' package may receive some
adjustments by the time a tutorial is written for it.  Following this
check-in, the repository will be considered stable, and CVS checkouts
of PEAK will be possible.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@432 571e12c6-e1fa-0310-aee7-ff1267fa46bd