New semantics for 'lookupComponent()': it's now fully integrated with peak.naming, so URLs and names acquired "off the top of the hierarchy" should now work.  It also takes '/'-separated paths now instead of '.'-separated ones, and includes support for absolute paths (beginning with '/') and relative ones ('.', '..').  The name is now a required argument (it used to be optional and would do the equivalent of '..').
In addition, there are many refinements to the binding package,
including lots of new module-level component navigation functions like
'getRootComponent()', 'acquireComponent()', etc.  See the docs for
details.  Almost all of them have tests, too.  Last, but not least,
I hacked on the naming API/SPI modules so that they'll work for now
with an empty default environment and using 'AbstractContext' as the
default 'InitialContext'.  That'll let us use whatever URL context
providers we want to put in before we get around to implementing the
configuration system and its defaults...


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@437 571e12c6-e1fa-0310-aee7-ff1267fa46bd