Properties and utilities are unified at last!  Both properties and utilities are now handled by the __instance_provides__ and __class_provides__ attributes of component classes.  This simplified the code and interfaces for both PropertyMap and for registering dynamic utility providers.
Property names now have an additional syntax form: 'some.prop?' which is
used as a default value that is checked after 'some.prop', 'some.*', and
'some.prop' (a second time).  This is handled transparently for you by
way of the 'setDefault()' method of 'IPropertyMap', so effectively it's an
implementation detail.

Documentation is still a little sketchy, but it should now be
straightforward to get the naming system using properties and utilities in
a harmonious way.


ALSO:

Fixed config.getLocal() to ensure it's working with a root component.
binding.New() can now use an import string instead of a
class/type/callable.  The object referenced by the import string will be
loaded and called as though the object itself had been used.  GlobalConfig
and LocalConfig now have 'setup()' methods that can be overridden to
create default rule and provider registrations.  This'll be handy for
making LocalConfig register a default transaction service provider, for
example.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@555 571e12c6-e1fa-0310-aee7-ff1267fa46bd