Added a basic configuration file format, and used it to replace the schemes dictionary in peak.naming.factories, by adding a "peak.ini" global configuration file for PEAK builtins.  This means that the factories subpackage is pretty much useless now; the URL classes can easily move anywhere now, like to the same modules as the implementations. It should now be easy to add lots of address schemes to PEAK.  Also, subclasses of GlobalConfig can change the list of filenames used to load configuration settings, and thus perhaps read a ~.peakrc or /etc/peak.ini file, or an app-specific global configuration file.
The file format is very primitive right now: ConfigParser with eval() of
the settings done at file load time.  Ultimately, this should move to a
true lazy evaluation mode, and there are some syntactic sugar features that
should be added.  For right now, however, all we need are constant
expressions.

Last, but not least, fixed a bug in the fromURL() method of ParsedURLs.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@576 571e12c6-e1fa-0310-aee7-ff1267fa46bd