Refactored so that config.loadConfigFile() and config.loadMapping() can be called on any IConfigurable - which means any Component.  So now you can do things like 'config.loadConfigFile(config.getGlobal(), somefile)' or 'config.loadMapping(myComponent, config.fileNearModule('foo','bar.ini'))'.
Also fixed an issue where PropertyMap objects wouldn't overwrite an
existing provider if the existing one was registered under the same
interface or property name.

Last, but not least, re-arranged and enhanced both binding.interfaces and
config.interfaces to reflect the nature of things that supply
configuration data, things which can be configured, and the ways in which
these things are accomplished.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@613 571e12c6-e1fa-0310-aee7-ff1267fa46bd