Removing excessive DWIMishness from peak.naming, part 1:
Made ParsedURLs accept a scheme and body, instead of trying to guess from
a string what they should do.  It's now the *sole* responsibility of
naming.toName() to determine whether a string should be treated as a URL,
and then only if it's told to make that determination.  Next up: refactor
naming context base classes to have unambiguous rules for how they parse
and process names.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@642 571e12c6-e1fa-0310-aee7-ff1267fa46bd