Tentative SQL typemap support
It's going to be slower than the old way of returning rows, since it's
going to call something on each field of each row.  :(  I suppose I could
add code to skip the typemap work if all the converters are
'NullConverter', but such a condition seems too rare to be worth trying
to optimize for.  Most likely, if we have speed issues, we'll want to make
the converters themselves fast.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@662 571e12c6-e1fa-0310-aee7-ff1267fa46bd