Small speedup (about 3/4ths of a second on my PC) of UML metamodel load.
XMIMapMaker was doing a lot of unnecessary work to create the XMI map; of
the 9.5 seconds the profiler claims is UML model loading, 1.6 were spent
in XMIMapMaker.__init__!  Most of the remaining slowness is in calls to
'prepForSimulation()', so any significant further speedups will have to
come from translation of that and 'peak.utils.Code' to C.  Strangely, none
of this stuff is anywhere near as slow on Unix machines as it is on my
Windows PC.  So I guess we can wait before further optimizing.  But this
was a quick win, at least for the UMLModel test suite.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@693 571e12c6-e1fa-0310-aee7-ff1267fa46bd