Squeezed out another ~100ms by removing support for passing opcode names to certain functions; they now require numeric values.  It's amazing how much of a difference it can make to remove one 'isinstance()' check, at least if it's being called over 18000 times.  :)  On the other hand, it's kind of amazing that doing an 'isinstance()' check 18000 times only takes 100ms!
I think this is going to be the last module inheritance optimization for a
while again.  It's now fast enough for me to begin some serious work on
XMI and UML models without experiencing excess pain during the debugging
process waiting for the code to finish loading.  :)  Guess we'll see how it
goes from here.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@764 571e12c6-e1fa-0310-aee7-ff1267fa46bd