Added exceptions for when a context in the terminal naming system does not support the required interface.  I think this "resolves" the last issues with supporting the "resolve" operation; it's now up to the implementors of contexts to do their thing.
One thing to consider, though...  should we make 'nameClass' default to
'CompositeName' now?  For flat namespaces, this would mean they'd get
automatic '/' support, so the NIS stuff wouldn't need to have specific
code for that, for example.  I'll have to see what Ty thinks and also
double-check that it won't break anything for URL contexts, etc.  I'd also
have to tweak _resolveComposite() so that if 'nameClass' is 'Composite'
that it wraps the recursive name call in a 'CompoundName'.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@777 571e12c6-e1fa-0310-aee7-ff1267fa46bd