Beginning of refactoring to support naming authorities, absolute addresses, etc.  This is just a few small fixups; the main changes will be coming later in the refactoring to make "Address = Parameters + Name", and to move most of the "name arithmetic" into the name and address classes.  It should then end up pretty easy to do most of the things people will want to do with new naming or address systems.
One of the first changes here is to make AbstractContext -> NameContext,
and GenericURLContext -> AddressContext.  Nameless, contextless addresses
will use AddressContext, and anything which contains a name will implement
an address class, and a context based on NameContext.  All of these names
may change before the refactoring is done, which is why I've left in
aliases to the old names for now.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@779 571e12c6-e1fa-0310-aee7-ff1267fa46bd