Hierarchical contexts and composite namespaces work now.  I had to refactor some of the context base class internals to get this to work reasonably, and there are now some new methods that can be overridden.  In the simple case, overriding '_contextNNS()' lets you handle per-context NNS pointers. (And in the really simple case, you can ignore it altogether.)  If you want to support explicit NNS pointers, you can override _get_nns(), _bind_nns(), and _unbind_nns() as well.
The rules of name arithmetic have changed slightly; any addition that
previously returned a composite name will now return a compound name, if it
is possible to do so without changing the meaning of the returned name.
Similarly, the parser for nesting a compound name inside a composite name,
will return just a compound name if possible/appropriate.

There are new 'naming.isBoundary()' and 'naming.crossesBoundary()' APIs
that can be used to check if a name is a terminal namespace boundary, or
crosses namespace boundaries, respectively.

The name parsing tests have been updated for arithmetic and parsing
changes, and also to use the new naming.parseURL() function instead of
performing partial retrievals on the names.

Finally, a new 'config:' URL scheme has been added, as a demo of the
finished hierarchy and composite name support.  For example,
'config:environ.TMP' refers to a 'PropertyContext()' for that property
namespace, but 'config:environ.TMP/' returns the value of 'environ.TMP'.
Similarly, 'naming.lookup("config:environ")["TMP/"]' returns the value,
while 'naming.lookup("config:environ")["TMP"]' is just another naming
context.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@790 571e12c6-e1fa-0310-aee7-ff1267fa46bd