Partial refactoring of peak.model class hierarchy.  XMI support has been moved to an abstract 'model.Namespace' class that 'model.Package' and 'model.Classifier' inherit from.  A separate Classifier subclass, 'model.Immutable', is now used to add immutability features.
This is barely the beginning of the refactoring, however.  A lot of
metadata class attributes need to be added to 'model.Namespace', and the
feature classes need to be cleaned up to work more on metadata and less
on hardcoded methods.  More relationship metadata is also needed, like
composition/aggregation info.  Features should be able to compute their
own XMI names based on package containment, and packages should be able
to find classes nested in inner packages (to support short-name lookups for
XMI 1.1).

There should also be a module of "primitive type" implementations for
things like Integer and String.  The parsing/formatting/conversion
signatures aren't entirely clear yet, though.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@802 571e12c6-e1fa-0310-aee7-ff1267fa46bd