Beginning implementation of a full MOF 1.3.1 metamodel.  We'll need this to process current/valid OMG metamodels for UML, CWM, etc.  Plus, it'll be a good basis for generating Python code.  UML and other high-level design metamodels are too abstract for easy code generation, but MOF is designed to be a very concrete representation of a program's implementation model. We'll generate MOF models from UML models before generating artifacts such as Python code or DDL.  MOF is also the basis of XMI, so in general it should be useful to have around.
Unfortunately, the 600+ lines of code I've written so far only fully
implement maybe 2-3 of the 23 MOF classes, just sketching the inheritance
hierarchy and contents rules for the rest.  I've got about another 50 pages
of MOF spec to implement yet, and some of what I've done I still need to
go back and add a bit into 'peak.model' to support properly.  :(  The good
news is that this is a good test of 'peak.model', and it'll help us to get
through the coming refactoring of 'peak.model'.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@812 571e12c6-e1fa-0310-aee7-ff1267fa46bd