Refactored bidirectional association protocol, and data manipulation SPI. Specifically, the '_notifyLink()' and '_notifyUnlink()' methods of feature objects are always called to perform any modification on the primary side (i.e. the side whose API was called) and '_link()' and '_unlink()' are now always called to perform any actual modification of the reference's list. (This required the addition of an optional "position" parameter to these methods.)  This should make it easier to incorporate things like observers, and "specialty" feature types like qualified associations or mappings.
One possible consequence of this refactoring could be reduced performance,
as there are more method calls taking place for many activities.  But in
most applications that manipulate association links, the overhead is
probably negligible compared to the mechanics of persisting those changes
anyway.  (These changes only affect modifications, and only to
association ends.)


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@845 571e12c6-e1fa-0310-aee7-ff1267fa46bd