Major refactoring of 'peak.model'.  Changes include:
* Added 'enumerations' module with the new implementation of
  'model.Enumeration' -- read its (thorough) docs for details on the new
  features.  Changed existing uses/test cases for the new "spelling"
  of enumerations.

* The 'interfaces' module has been thoroughly revised to drop ancient
  TransWarp stuff, and to mostly match the actual structure of today's
  'peak.model' package.  It still needs to be gone over for errors and
  omissions, and the docs need to be fleshed out some more, but it
  should actually provide some useful information now.

* Features now support an 'implAttr' attribute which can specify what
  attribute is actually used to store the feature.  It defaults to being
  the same as the feature's name, but you can override it in the feature
  definition to redirect its storage to a private attribute.

* There is now a 'structField' feature type that can be used to define
  read-only fields for immutable types without having to keep typing
  'isChangeable = False'.

* Features now won't actually allow themselves to be changed, if their
  'isChangeable' flag is false.

* Many methods have moved around, in an attempt to consolidate the
  class hierarchy for features.  This process may continue for a bit.

This check-in completes about 60-80% of the 'peak.model' refactoring that
I announced on the mailing list.  It has now progressed far enough that I
can begin experimenting with the MOF and XMI features that were blocked
waiting for this work.  Implementing those features may drive further
changes to the 'peak.model' package, but I believe the API of
'peak.model' is no longer dangerously unstable.  (One known change
that's still to come: default constructor implementations for
'model.Immutable' and 'model.Element'.)

Prior to this refactoring, it was unclear what paradigms would come out
on top in certain areas, and that uncertainty is now mostly eliminated.
It's most likely that further refactorings in 'peak.model' will be mere
"spelling" or class hierarchy implementation issues, rather than
fundamental changes of approach.  But I'm going to wait until XMI
writing and MOF->Python code generation are further along before
declaring the API "stable" on the general mailing list.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@849 571e12c6-e1fa-0310-aee7-ff1267fa46bd