Refactoring to support uniform internal data structure for features.  All features with a maximum multiplicity of 1 are now stored simply as the object representing their value.  (Previously, 'model.Reference' features were represented as a list containing zero or one items.)  This should make it easier for most people implementing data managers to get the object state correct, but it was necessary to make some changes to the current XMI DM to support it.
This change is part of a general trend towards encapsulating all of
features' behavior in one or two base classes, with the current Reference,
Field, Sequence, etc. classes becoming just shorthand for specifying the
appropriate metadata.  (This will ultimately make it easier to do code
generation that's based on pure metadata from a MOF or UML model.)


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@851 571e12c6-e1fa-0310-aee7-ff1267fa46bd