More TypeCode bootstrapping.  Enumeration and DataType can now generate their own TypeCode values (as 'mdl_typeCode'), Primitive defaults to 'tk_any', and all other Classifier subclasses default to a type kind of 'tk_objref'.  Of course, to be useful, primitive types should be set to an appropriate CORBA typecode for efficient and type-safe XMI rendering. Also, made TypeCodes repr()-able for debugging, and did some more cleanup to catch MOF131 up to the current state of peak.model.
I think maybe we should add some additional primitive datatypes for ease of
model definition; perhaps such that 'String(30)' actually gives you an
object whose mdl_typeCode is 'TypeCode(kind=TCKind.tk_string,length=30)'.
That's a somewhat radical interpretation of what a PrimitiveType is,
however.  I'll need to give it some thought.

I've also been thinking that it would be nice to be able to put some XMI
helper code into the peak.model base classes.  But the problem with that is
that it's too likely to be dependent on XMI versions, and will scatter the
XMI implementation all over creation.  This is definitely the sort of thing
AOP was intended to solve, but frankly PEAK's AOP capabilities suck.  I
think that instead, some kind of adaptation mechanism is needed.  I wonder
if I could create some kind of AdapterDM for this purpose?


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@865 571e12c6-e1fa-0310-aee7-ff1267fa46bd