Change to binding.Once: if the object's dict contains 'NOT_FOUND' as the bound attribute's value, 'AttributeError' is raised upon access.  This lets you override a bound attribute value and declare it non-existent, and it also ensures that a circularly defined attribute will result in an error message.
Unfortunately, an error will also result if two threads attempt
to access and compute the bound attribute "at the same time".  But this
is a slight improvement over the situation prior to this change, where a
race condition would cause the "losing" thread to silently return
'NOT_FOUND', possibly causing the problem to lurk awhile.

The existence of this race condition is a potentially serious problem with
the use of lazy attribute bindings in class objects (which are shared
between threads), and it may be that we will need to require some type of
locking for such bindings.  :(


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@917 571e12c6-e1fa-0310-aee7-ff1267fa46bd