Made Once attributes thread-safe, to a point.  While a "once" attribute is being calculated, a special "lock" object is placed in the object's dictionary.  This object just holds the thread ID of the calculating thread.  If a re-entrant access to the attribute is made, the lock's ID is checked for whether it's the same thread.  If it's the same thread, an error is raised indicating that the attribute has a circular definition. If it's not the same thread, the current thread pretends the value hasn't been computed yet, and tries to compute it.
Note that in cases where the underlying once function and object are
properly constructed, this may occasionally result in two threads doing
the same work to calculate the attribute, but the results are always
the same.  In cases where the once function actually refers back to itself,
it's possible for competing threads to keep stealing the lock from each
other and thus never terminate.  Well, actually, you'll get a recursion
depth error on both threads, eventually.  So the limitations of this guard
technique should be quite acceptable for the intended uses of once
bindings, especially since this protection is really intended for class
attributes such as 'peak.model' metadata, not instance attributes.  Classes
*have* to be sharable between threads; instances rarely need to be.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@926 571e12c6-e1fa-0310-aee7-ff1267fa46bd