Added support for XMI DM's to look up metamodels via properties.  The 'peak.xmi.metamodels' property namespace will be checked for a property named 'name.version' where 'name' is the 'xmi.name' attribute of the 'XMI.metamodel' tag, and 'version' is the 'xmi.version' attribute, if present.  The property value should be the appropriate metamodel module.
The 'DM.importFromXMI()' method has been replaced with a module-level
function, 'storage.xmi.fromFile()'.  For most simple XMI-processing
programs, simply calling the function to get back the top-level content
list for a specified file or URL should be sufficient.  It's still
possible, however, to manually specify a metamodel, using the 'metamodel'
keyword argument to 'fromFile()'.

In addition, this checkin contains some changes I previously made to
simplify the "peak.ini" using the new 'config.instancePerApp()' API,
but accidentally skipped checking in.  Last, but not least, there are
some changes to 'peak.model.features' that make it easier to introspect
a feature's CORBA typecode kind (which is important for XMI writing), and
I stripped trailing whitespaces from the source of the
'peak.storage.data_managers' module.  (I should be stripping them from
many more modules too, as soon as I figure out how to automate it on a
multi-file basis in jEdit.)


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@952 571e12c6-e1fa-0310-aee7-ff1267fa46bd