First draft implementation of system scheduling proposal.  (Also, misc. TODO additions.) Interfaces have been streamlined and names changed. Daemons are now "periodic tasks", and the provided base class implements an "adaptive task" interface.  "Daemonic app", "event loop", and "daemon scheduler" are now just a single "system scheduler" interface, which can schedule periodic tasks as well as one-offs.  I settled on "higher priority number" = "higher priority" for priority sorting.
I'm beginning to like this whole "asynchronous programming" model that
Twisted provides.  AdaptiveTask (formerly BaseDaemon) is a lot simpler,
MultiDaemon went away, and the system scheduler is only two screens
of code to manage idle timeouts, prioritized task scheduling, and more.
Heck, "shutdown after X" was as simple as 'callLater(duration,self.stop)'!

There is one minor problem in this draft...  if a task raises an unhandled
exception, it won't be scheduled for further re-run.  I'm not sure if this
is a bug or feature yet.  It should be easy to fix, though.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@982 571e12c6-e1fa-0310-aee7-ff1267fa46bd