Adjustments to scheduling algorithms.  Tasks are now more sensitive to their priority, to better match MetaDaemon "greedy" approach.  Before this adjustment, only tasks that were behind schedule for execution were prioritized, and they were run one after another with no opportunity for high priority tasks to "jump back to the head of the line".  Now, if a high priority task is rescheduled before an earlier-scheduled low priority task, it *will* jump to the head of the line, and thus starve lower priority tasks until it becomes idle again.
The other adjustment made was to fix a problem where an idle timeout might
not be detected until the system had been idle for up to twice the timeout.
The idleness-checker now schedules itself to rerun at the earliest possible
moment of timeout, instead of waiting another full idle checking interval.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@983 571e12c6-e1fa-0310-aee7-ff1267fa46bd