Finished the bulk of the "daemon" refactoring.  There are now separate interfaces for different aspects of task scheduling and event loop management.  An 'ITaskQueue' schedules prioritized periodic tasks, an 'IMainLoop' controls activity monitoring and timeouts, and an 'IBasicReactor' or 'ITwistedReactor' handles I/O and scheduled event processing.  There are default implementations of all scheduling interfaces (including an untested 'UntwistedReactor' for use when Twisted isn't needed or available), and they are registered in 'peak.ini' as utility providers.
In addition, I added two application-related interfaces, 'ICmdLineApp' and
'ICmdLineAppFactory'.  These are intended to form the basis of a framework
for handling CGI, FastCGI, or any other command-line invocation protocol
such as ReadyExec.  That is, a protocol handler for (say) FastCGI would
invoke an 'ICmdLineAppFactory' to create an 'ICmdLineApp' with a given
stdin, stdout, stderr, environ, and argv, ready to be 'run()'.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@991 571e12c6-e1fa-0310-aee7-ff1267fa46bd