More "root" cleanups.  'config.makeRoot()' can now take an 'iniFiles' argument listing file paths (or '(modulename,filename)' tuples) to load; it defaults to '[('peak','peak.ini')]', providing the default behavior we know and love. :)
If you create a binding.Component with a parent of 'None', it will now
believe you and make itself a root.  The default value of parentComponent
is now 'NOT_GIVEN', so that if you don't specify a parent, the object will
be "undecided" until it really needs to know who its parent is (at which
point it will be 'None' if it hasn't been "adopted" in the meantime).
Interface docs and tests have been updated to reflect/take advantage of
this.  (Specifically, the tests used to use 'getParentComponent()' to force
a root to acknowledge its orphan status; now they can just *tell* it.)

Finally, I removed the use of weak references from the assembly events
system.  If you created an object, intending that it register itself with
some kind of registry, it could get GC'd before the assembly event kicked
in, unless you specifically held onto a reference until then.  Since
components that don't need assembly events don't register for them, it
seems harmless to use simple references rather than weak ones.


git-svn-id: svn+ssh://svn.eby-sarna.com/svnroot/PEAK@1054 571e12c6-e1fa-0310-aee7-ff1267fa46bd